# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Zen-Live API",
    "description": "Auto-generated Swagger docs",
    "version": "1.0.0"
  },
  "host": "api.zen-urbancode.in",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT via Bearer scheme. e.g. \"Authorization: Bearer <token>\""
    }
  },
  "paths": {
...
    "/leads/form-submit": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "mobile_number": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                },
                "college_company": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                },
                "source": {
                  "example": "any"
                },
                "course_type": {
                  "example": "any"
                },
                "course_name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/trainer/leads": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/trainer/leads/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "mobile_number": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                },
                "college_company": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                },
                "source": {
                  "example": "any"
                },
                "course_type": {
                  "example": "any"
                },
                "course": {
                  "example": "any"
                },
                "batch_id": {
                  "example": "any"
                },
                "trainer_name": {
                  "example": "any"
                },
                "trainer_mobile": {
                  "example": "any"
                },
                "trainer_email": {
                  "example": "any"
                },
                "comments": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "paid_status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/filter-trainers": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}